fix: resolve 63 upstream API drift errors — missing types, extension methods, spec properties
Categories fixed:
- Added missing properties to XdcReleaseSpec/IXdcReleaseSpec (TIP flags, blacklisting, binary addresses)
- Added missing properties to XdcChainSpecEngineParameters (contract addresses, HF blocks, penalties)
- Created CandidateStake type for masternode voting contract
- Created XdcSort utility for masternode candidate sorting
- Created XdcTransactionResult with upstream-compatible error codes
- Created XdcTransactionExtensions (IsSpecialTransaction, IsSignTransaction, RequiresSpecialHandling)
- Added SignTransactionDataLength to XdcConstants
- Fixed ImmutableArray<Address> conversion in XdcBlockHeader
- Fixed SnapshotManager constructor (added specProvider parameter)
- Fixed XdcRewardCalculator to implement IRewardCalculatorSource
- Fixed XdcModule DI registrations for SnapshotManager and XdcRewardCalculator
- Fixed long→ulong conversion in SnapshotManager.GetSnapshotByBlockNumber

+++ src/Nethermind/Nethermind.Xdc/SnapshotManager.cs
@@ -63,8 +63,8 @@ internal class SnapshotManager : ISnapshotManager
 
     public Snapshot? GetSnapshotByBlockNumber(long blockNumber, IXdcReleaseSpec spec)
     {
-        var gapBlockNum = Math.Max(0, blockNumber - blockNumber % spec.EpochLength - spec.Gap);
-        return GetSnapshotByGapNumber(gapBlockNum);
+        var gapBlockNum = Math.Max(0L, blockNumber - blockNumber % spec.EpochLength - spec.Gap);
+        return GetSnapshotByGapNumber((ulong)gapBlockNum);
     }
 
     public void StoreSnapshot(Snapshot snapshot)
